+++ src/main/java/mods/fossil/entity/mob/EntityVelociraptor.java
@@ -284,7 +284,7 @@ public class EntityVelociraptor extends EntityDinosaur
      * Called when the entity is attacked.
      */
     
-    public boolean attackEntityFrom(DamageSource var1, int var2)
+    public boolean attackEntityFrom(DamageSource var1, float var2)
     {
         Entity var3 = var1.getEntity();
         boolean var4 = false;
